cleaning

+++ classes/PDFFilter.php
@@ -7,18 +7,24 @@ class PDFFilter
     $this->accountId=$accountId;
   }
 
-  public function filter($url )
+  public function filter($document)
   {
-    $content = file_get_contents($url);
-    $tmpFile=$this->accountId."tmp";
-    unlink($tmpFile.".pdf");
-    unlink($tmpFile.".txt");
-    $fh = fopen($tmpFile.".pdf",'w');
-    fwrite($fh, $content);
-    fclose($fh);
-    system("pdftotext ".$tmpFile.".pdf");
-    $txt = file_get_contents($tmpFile.".txt");
-    return($txt);
+    if(strlen($document->content)>0){
+      print "pdf - start - [".$document->url."] \r\n";
+      $tmpFile=$this->accountId."tmp";
+      unlink($tmpFile.".pdf");
+      unlink($tmpFile.".txt");
+      $fh = fopen($tmpFile.".pdf",'w');
+      fwrite($fh, $document->content);
+      fclose($fh);
+      system("pdftotext ".$tmpFile.".pdf");
+      $txt = file_get_contents($tmpFile.".txt");
+      print $txt; 
+      print "pdf - end - [".$document->url."] \r\n";
+      return($txt);
+    } else {
+      print "pdf - no content - [".$document->url."] \r\n";
+    } 
   }
 };
 ?>
